Solve model adjustment

+++ cubetimer_app/models.py
@@ -2,7 +2,7 @@ from django.db import models
 from django.contrib.auth.models import User
 
 class Solve(models.Model):
-    solvetime = models.IntegerField()
+    solvetime = models.FloatField()
     event = models.CharField(max_length=50)
     date = models.DateTimeField(auto_now_add=True)
     solved_by = models.ForeignKey(User, on_delete=models.CASCADE, related_name='solves')
